Add LED Type and Prepare for WS8211 Support
[Problem]
As we switch between different LED types, there needs
to be a way to select the LED type. Also, we need to explicitly
enable the 5v PIN when not plugged into USB.

[Solution]
Added an option for LED type selection, minor cleanups to the way
pins are selected, and setup to enable and disable the 5v power
pin appropriately.

[Testing]
Confirmed working on device.